+++ osquery/core/CMakeLists.txt
@@ -95,6 +95,7 @@ function(generateOsqueryCore)
   add_test(NAME osquery_core_tests_tablestests-test COMMAND osquery_core_tests_tablestests-test)
   add_test(NAME osquery_core_tests_watcherpermissionstests-test COMMAND osquery_core_tests_watcherpermissionstests-test)
   add_test(NAME osquery_core_tests_querytests-test COMMAND osquery_core_tests_querytests-test)
+  add_test(NAME osquery_core_tests_processtests-test COMMAND osquery_core_tests_processtests-test)
 
 endfunction()
